메서드 명 수정, 달성도 DTO 추가

+++ src/main/java/com/team/routineconnect/repository/DayOrderRepositoryCustom.java
@@ -14,5 +14,5 @@ public interface DayOrderRepositoryCustom {
     Optional<Float> findMaxPositionByUserAndDate(User user, LocalDate date);
     List<LocalDate> findDatesByUserAndDayAndDayGreaterThan(User user, DayOfWeek day, LocalDate date);
     List<RoutineWithAccomplishment> findRoutinesByUserAndDate(User user, LocalDate date);
-    Double findAchievementByUserAndDate(User user, LocalDate date);
+    Float findAchievementByUserAndDate(User user, LocalDate date);
 }
